+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added `ydb.WithCompression()` option
 * Added `ydb.MustOpen` and `ydb.MustConnector` helpers
 * Fixed `internal/xerrors.Transport` error wrapping for case when given error is not transport error
 * Added `grpc.WithBlock()` dial option to defaults